common: give our AWS fixtures a consistent behaviour
